Move all the components to the features folder
