+++ upflux-client/src/App.tsx
@@ -1,11 +1,11 @@
 import logo from "./assets/logos/logo-dark.png";
 import './App.css';
 import { BrowserRouter as Router } from 'react-router-dom'; 
-import Navbar from './components/Navbar'; 
-import HeroImage from './components/Header'; 
-import About from './components/About';
-import ContactUs from './components/ContactUs';
-import Footer from './components/Footer';
+import Navbar from './features/Navbar/Navbar'; 
+import HeroImage from './features/Header/Header'; 
+import About from './features/About/About';
+import ContactUs from './features/ContactUs/ContactUs';
+import Footer from './features/Footer/Footer';
 
 
 export const App = () => {
